refact/front/ButtonGreen.jsx: moves to src/components/buttons/ButtonGreen.jsx
