+++ frontend/src/pages/trips/ReserveTrip.jsx
@@ -5,7 +5,7 @@ import { setNavigationLayout } from "../../store/slices/navigationSlice"
 import DateDisplay from "../../components/DateDisplay"
 import Time from "../../components/Time"
 import LocationBarIcons from "../../components/LocationBarIcons"
-import ButtonGreen from "../../components/ButtonGreen"
+import ButtonGreen from "../../components/buttons/ButtonGreen"
 import { useState } from "react"
 import PopupReserveTrip from "../../components/PopupReserveTrip"
 import Switch from "../../components/Switch"
